Adding a 64 bit build flag and fixing a 64 bit build problem.

+++ app/audioscrobbler/MediaDevices/DeviceScrobbler.cpp
@@ -106,7 +106,7 @@ DeviceScrobbler::iPodDetected( const QStringList& arguments )
     
     if( pos > -1 ) serialNumber = arguments[ pos + 1 ];
    
-    qDebug() << "emitting detectedIPod: " << (int)this;
+    qDebug() << "emitting detectedIPod: " << (long)this;
     emit detectedIPod( serialNumber );
 }
